@@ -13,8 +13,6 @@ classifiers =
     License :: OSI Approved :: BSD License
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     Programming Language :: Python :: 3.12
@@ -24,7 +22,7 @@ packages = find:
 install_requires =
     bitstring
     cffi>=1.0.3;implementation_name == 'cpython'
-python_requires = >=3.8
+python_requires = >=3.10
 include_package_data = True
 
 [options.extras_require]
